Handle nested models.
