+++ betfair/meta/types.py
@@ -1,11 +1,29 @@
 # -*- coding: utf-8 -*-
 
-from schematics.types import BaseType
+from schematics import types
+from schematics.types import compound
 from schematics.exceptions import ConversionError
 from schematics.exceptions import ValidationError
 
+from . import utils
 
-class EnumType(BaseType):
+
+class DateTimeType(types.DateTimeType):
+
+    DEFAULT_FORMATS = (
+        '%Y-%m-%dT%H:%M:%S.%f',
+        '%Y-%m-%dT%H:%M:%S.%fZ',
+        '%Y-%m-%dT%H:%M:%S',
+    )
+
+
+class ModelType(compound.ModelType):
+
+    def export_loop(self, *args, **kwargs):
+        return utils.serialize_dict(super(ModelType, self).export_loop(*args, **kwargs))
+
+
+class EnumType(types.BaseType):
 
     MESSAGES = {'choices': u'Value must belong to enum {0}.'}
