+++ betfair/models.py
@@ -6,12 +6,12 @@ from schematics.types import IntType
 from schematics.types import FloatType
 from schematics.types import StringType
 from schematics.types import BooleanType
-from schematics.types import DateTimeType
 from schematics.types.compound import DictType
 from schematics.types.compound import ListType
-from schematics.types.compound import ModelType
 
 from betfair.meta.types import EnumType
+from betfair.meta.types import ModelType
+from betfair.meta.types import DateTimeType
 from betfair.meta.models import BetfairModel
 
 from betfair import constants
